+++ backend/src/test/java/com/autotrade/dashboard/indicator/IndicatorControllerTest.java
@@ -39,7 +39,8 @@ class IndicatorControllerTest {
     @Test
     void indicators_registeredTicker_returns200() throws Exception {
         Ticker ticker = new Ticker("AAPL", AssetType.STOCK, "NASDAQ");
-        MacdResult macd = new MacdResult(new BigDecimal("2.11694333"), new BigDecimal("2.13097767"), new BigDecimal("-0.01403434"));
+        MacdResult macd = new MacdResult(new BigDecimal("2.11694333"), new BigDecimal("2.13097767"),
+                new BigDecimal("-0.01403434"), new BigDecimal("0.0124"));
         MovingAverageResult ma = new MovingAverageResult(10, new BigDecimal("111.92000000"), 30,
                 new BigDecimal("108.94000000"), MovingAverageRelation.SHORT_ABOVE_LONG);
         IndicatorResponse response = IndicatorResponse.from(ticker, Broker.ALPACA,